+++ resources/api/global.d.ts
@@ -53,16 +53,30 @@ declare global {
     }
 
     /**
-     * Namespace providing window / UI functionality.
+     * Namespace providing command APIs.
+     */
+    readonly commands: {
+      /**
+       * Executes one or more commands in sequence.
+       * @param cmds Each command may be specified as either:
+       * - a string command ID
+       * - a `[id, ...args]` tuple providing the command ID and its arguments.
+       * @returns A promise that resolves after all commands have completed.
+       */
+      executeCommands(...cmds: Array<string | [id: string, ...args: Array<any>]>): Promise<void>;
+    }
+
+    /**
+     * Namespace providing window and UI APIs.
      */
     readonly window: {
       /**
-       * Assigns an available TreeView ID to the caller.
+       * Returns an available TreeView ID for the caller.
        */
       getTreeViewId(): string | undefined;
 
       /**
-       * Assigns an available WebView ID to the caller.
+       * Returns an available Webview ID for the caller.
        */
       getWebviewId(): string | undefined;
 
@@ -72,7 +86,7 @@ declare global {
       releaseTreeViewId(id: string): boolean;
 
       /**
-       * Releases a previously assigned WebView ID.
+       * Releases a previously assigned Webview ID.
        */
       releaseWebviewId(id: string): boolean;
     }