+++ resources/templates/js/command.js
@@ -1,17 +1,11 @@
 
 const { userInfo } = require('os');
 
-async function runCommands(...cmds) {
-  for (const { cmdId, args = [] } of cmds) {
-    await vscode.commands.executeCommand(cmdId, ...args);
-  }
-}
-
 // Insert a TODO comment at current cursor line.
 // Reference: https://code.visualstudio.com/api/references/commands
-runCommands(
-  { cmdId: 'editor.action.insertLineBefore' },
-  { cmdId: 'type', args: [{ text: `TODO (${userInfo().username}): <describe task>` }] },
-  { cmdId: 'editor.action.addCommentLine' },
-  { cmdId: 'cursorEnd' },
+macros.commands.executeCommands(
+  'editor.action.insertLineBefore',
+  ['type', { text: `TODO (${userInfo().username}): <describe task>` }],
+  'editor.action.addCommentLine',
+  'cursorEnd',
 );